# apps/backend/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Aspiron API",
    "description": "Backend API for the Aspiron project",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/api/v1/attempts": {
      "parameters": [
        {
          "name": "x-client-type",
          "in": "header",
          "description": "Client type header",
          "required": false,
          "example": "BROWSER"
        }
      ],
      "post": {
        "tags": [
          "Assessment"
        ],
        "operationId": "handler_create_attempt",
        "responses": {
          "200": {
            "description": "Create assessment attempt"
          }
        }
      }
    },
    "/api/v1/attempts/{attempt_id}/results": {
      "parameters": [
        {
          "name": "x-client-type",
          "in": "header",
          "description": "Client type header",
          "required": false,
          "example": "BROWSER"
        }
      ],
      "get": {
        "tags": [
          "Assessment"
        ],
        "operationId": "handler_get_results_by_attempt_id",
        "parameters": [
          {
            "name": "attempt_id",
            "in": "path",
            "description": "Attempt ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get attempt results"
          }
        }
      }
    },
    "/api/v1/attempts/{attempt_id}/submit": {
      "parameters": [
        {
          "name": "x-client-type",
          "in": "header",
          "description": "Client type header",
          "required": false,
          "example": "BROWSER"
        }
      ],
      "post": {
        "tags": [
          "Assessment"
        ],
        "operationId": "handler_submit_attempt",
        "parameters": [
          {
            "name": "attempt_id",
            "in": "path",
            "description": "Attempt ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Submit assessment attempt"
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "parameters": [
        {
          "name": "x-client-type",
          "in": "header",
          "description": "Client type header",
          "required": false,
          "example": "BROWSER"
        }
      ],
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "authenticate_user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "User authenticated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_MobileTokenResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "parameters": [
        {
          "name": "x-client-type",
          "in": "header",
          "description": "Client type header",
          "required": false,
          "example": "BROWSER"
        }
      ],
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "log_out",
        "responses": {
          "200": {
            "description": "User logged out"
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "parameters": [
        {
          "name": "x-client-type",
          "in": "header",
          "description": "Client type header",
          "required": false,
          "example": "BROWSER"
        }
      ],
      "get": {
        "tags": [
          "Auth"
        ],
        "operationId": "get_me",
        "responses": {
          "200": {
            "description": "Get current user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_AuthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/refresh-token": {
      "parameters": [
        {
          "name": "x-client-type",
          "in": "header",
          "description": "Client type header",
          "required": false,
          "example": "BROWSER"
        }
      ],
      "get": {
        "tags": [
          "Auth"
        ],
        "operationId": "refresh_token_handler",
        "responses": {
          "200": {
            "description": "Refresh access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_MobileTokenResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/register-user": {
      "parameters": [
        {
          "name": "x-client-type",
          "in": "header",
          "description": "Client type header",
          "required": false,
          "example": "BROWSER"
        }
      ],
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "register_user",
        "responses": {
          "200": {
            "description": "Register a new user"
          }
        }
      }
    },
...
    "/api/v1/quizzes/{quiz_id}": {
      "parameters": [
        {
          "name": "x-client-type",
          "in": "header",
          "description": "Client type header",
          "required": false,
          "example": "BROWSER"
        }
      ],
      "get": {
        "tags": [
          "Assessment"
        ],
        "operationId": "handler_fetch_quiz_by_id",
        "parameters": [
          {
            "name": "quiz_id",
            "in": "path",
            "description": "Quiz ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get quiz by ID"
          }
        }
      }
    },
    "/api/v1/quizzes/{quiz_id}/questions": {
      "parameters": [
        {
          "name": "x-client-type",
          "in": "header",
          "description": "Client type header",
          "required": false,
          "example": "BROWSER"
        }
      ],
      "get": {
        "tags": [
          "Assessment"
        ],
        "operationId": "handler_get_questions_by_quiz_id",
        "parameters": [
          {
            "name": "quiz_id",
            "in": "path",
            "description": "Quiz ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get questions by quiz ID"
          }
        }
      }
    },
...
    "/api/v1/topics/{topic_id}/quizzes": {
      "parameters": [
        {
          "name": "x-client-type",
          "in": "header",
          "description": "Client type header",
          "required": false,
          "example": "BROWSER"
        }
      ],
      "get": {
        "tags": [
          "Assessment"
        ],
        "operationId": "handler_get_quizzes_by_topic_id",
        "parameters": [
          {
            "name": "topic_id",
            "in": "path",
            "description": "Topic ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get quizzes by topic ID"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActionTypeEnum": {
        "type": "string",
        "enum": [
          "CREATE",
          "READ",
          "UPDATE",
          "DELETE",
          "MANAGE",
          "PUBLISH",
          "MODERATE",
          "GRADE",
          "TAKE",
          "VIEW_RESULTS",
          "VIEW_ANY_RESULTS",
          "ASSIGN_ROLES",
          "VIEW_ANALYTICS",
          "MANAGE_SETTINGS",
          "AUDIT"
        ]
      },
      "ApiResponse_AuthResponse": {
        "type": "object",
        "required": [
          "success"
        ],
        "properties": {
          "code": {
            "type": [
              "string",
              "null"
            ]
          },
          "data": {
            "type": "object",
            "required": [
              "user",
              "roles",
              "permissions"
            ],
            "properties": {
              "permissions": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/AuthPermissionResponse"
                }
              },
              "roles": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/AuthRoleResponse"
                }
              },
              "user": {
                "$ref": "#/components/schemas/AuthUserResponse"
              }
            }
          },
          "message": {
            "type": [
              "string",
              "null"
            ]
          },
          "success": {
            "type": "boolean"
          }
        }
      },
      "ApiResponse_MobileTokenResponse": {
        "type": "object",
        "required": [
          "success"
        ],
        "properties": {
          "code": {
            "type": [
              "string",
              "null"
            ]
          },
          "data": {
            "type": "object",
            "required": [
              "access_token",
              "refresh_token",
              "token_type",
              "expires_in"
            ],
            "properties": {
              "access_token": {
                "type": "string"
              },
              "expires_in": {
                "type": "integer",
                "format": "int64"
              },
              "refresh_token": {
                "type": "string"
              },
              "token_type": {
                "type": "string"
              }
            }
          },
          "message": {
            "type": [
              "string",
              "null"
            ]
          },
          "success": {
            "type": "boolean"
          }
        }
      },
      "AuthPermissionResponse": {
        "type": "object",
        "required": [
          "name",
          "resource_type",
          "action"
        ],
        "properties": {
          "action": {
            "$ref": "#/components/schemas/ActionTypeEnum"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "resource_type": {
            "$ref": "#/components/schemas/ResourceTypeEnum"
          }
        }
      },
...
      "AuthRoleResponse": {
        "type": "object",
        "required": [
          "name",
          "display_name",
          "is_system_role"
        ],
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "display_name": {
            "type": "string"
          },
          "is_system_role": {
            "type": "boolean"
          },
          "name": {
            "$ref": "#/components/schemas/UserTypeEnums"
          }
        }
      },
      "AuthUserResponse": {
        "type": "object",
        "required": [
          "id",
          "email",
          "is_active",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "avatar_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "created_at": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "first_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string"
          },
          "is_active": {
            "type": "boolean"
          },
          "last_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "updated_at": {
            "type": "string"
          }
        }
      },
...
      "LoginRequest": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "example": "barbara_turner.1@admin.aspiron"
          },
          "password": {
            "type": "string",
            "example": "admin123"
          }
        }
      },
...
      "ResourceTypeEnum": {
        "type": "string",
        "enum": [
          "USER",
          "CONTENT",
          "ASSESSMENT",
          "COMMUNITY",
          "SYSTEM",
          "SUBJECT",
          "CHAPTER",
          "TOPIC",
          "VIDEO",
          "QUIZ",
          "QUESTION",
          "THREAD",
          "POST",
          "NOTE"
        ]
      },
      "UserTypeEnums": {
        "type": "string",
        "enum": [
          "STUDENT",
          "TEACHER",
          "ADMIN"
        ]
      }
    }
  }
}